+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## NEXT
+## 0.3.0 [2023-10-04]
 
 - Adds placeholder login and sign up screens.
 - Use Firestore to store events.